// src/api-spec/protobuf/gen/js/tdex/v1/swap_pb.ts
export interface SwapRequest {
    /**
     * Random unique identifier for the current message
     *
     * @generated from protobuf field: string id = 1;
     */
    id: string;
    /**
     * The proposer's quantity
     *
     * @generated from protobuf field: uint64 amount_p = 2;
     */
    amountP: bigint;
    /**
     * The proposer's asset hash
     *
     * @generated from protobuf field: string asset_p = 3;
     */
    assetP: string;
    /**
     * The responder's quantity
     *
     * @generated from protobuf field: uint64 amount_r = 4;
     */
    amountR: bigint;
    /**
     * The responder's asset hash
     *
     * @generated from protobuf field: string asset_r = 5;
     */
    assetR: string;
    /**
     * The proposer's unsigned transaction in PSBT format (base64 string)
     *
     * @generated from protobuf field: string transaction = 6;
     */
    transaction: string;
    /**
     * In case of a confidential transaction the blinding key of each confidential
     * input is included. Each blinding key is identified by the prevout script
     * hex encoded.
     *
     * @generated from protobuf field: map<string, bytes> input_blinding_key = 7;
     */
    inputBlindingKey: {
        [key: string]: Uint8Array;
    };
    /**
     * In case of a confidential transaction the blinding key of each confidential
     * output is included. Each blinding key is identified by the output script
     * hex encoded.
     *
     * @generated from protobuf field: map<string, bytes> output_blinding_key = 8;
     */
    outputBlindingKey: {
        [key: string]: Uint8Array;
    };
}
/**
 * @generated from protobuf message tdex.v1.SwapAccept
 */
export interface SwapAccept {
    /**
     * Random unique identifier for the current message
     *
     * @generated from protobuf field: string id = 1;
     */
    id: string;
    /**
     * indetifier of the SwapRequest message
     *
     * @generated from protobuf field: string request_id = 2;
     */
    requestId: string;
    /**
     * The partial signed transaction base64 encoded containing the Responder's
     * signed inputs in a PSBT format
     *
     * @generated from protobuf field: string transaction = 3;
     */
    transaction: string;
    /**
     * In case of a confidential transaction the blinding key of each confidential
     * input is included. Each blinding key is identified by the prevout script
     * hex encoded.
     *
     * @generated from protobuf field: map<string, bytes> input_blinding_key = 4;
     */
    inputBlindingKey: {
        [key: string]: Uint8Array;
    };
    /**
     * In case of a confidential transaction the blinding key of each confidential
     * output is included. Each blinding key is identified by the output script
     * hex encoded.
     *
     * @generated from protobuf field: map<string, bytes> output_blinding_key = 5;
     */
    outputBlindingKey: {
        [key: string]: Uint8Array;
    };
}
/**
 * @generated from protobuf message tdex.v1.SwapComplete
 */
export interface SwapComplete {
    /**
     * Random unique identifier for the current message
     *
     * @generated from protobuf field: string id = 1;
     */
    id: string;
    /**
     * indetifier of the SwapAccept message
     *
     * @generated from protobuf field: string accept_id = 2;
     */
    acceptId: string;
    /**
     * The signed transaction base64 encoded containing the Proposers's signed
     * inputs in a PSBT format
     *
     * @generated from protobuf field: string transaction = 3;
     */
    transaction: string;
}
// ...
export const SwapRequest = new SwapRequest$Type();
// ...
export const SwapAccept = new SwapAccept$Type();
// @generated message type with reflection information, may provide speed optimized methods
class SwapComplete$Type extends MessageType<SwapComplete> {
    constructor() {
        super("tdex.v1.SwapComplete", [
            { no: 1, name: "id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "accept_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "transaction", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<SwapComplete>): SwapComplete {
        const message = { id: "", acceptId: "", transaction: "" };
        globalThis.Object.defineProperty(message, MESSAGE_TYPE, { enumerable: false, value: this });
        if (value !== undefined)
            reflectionMergePartial<SwapComplete>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: SwapComplete): SwapComplete {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string id */ 1:
                    message.id = reader.string();
                    break;
                case /* string accept_id */ 2:
                    message.acceptId = reader.string();
                    break;
                case /* string transaction */ 3:
                    message.transaction = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: SwapComplete, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string id = 1; */
        if (message.id !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.id);
        /* string accept_id = 2; */
        if (message.acceptId !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.acceptId);
        /* string transaction = 3; */
        if (message.transaction !== "")
            writer.tag(3, WireType.LengthDelimited).string(message.transaction);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message tdex.v1.SwapComplete
 */
export const SwapComplete = new SwapComplete$Type();