// src/api-spec/protobuf/gen/js/tdex/v1/types_pb.ts
export interface Fee {
    /**
     * @generated from protobuf field: int64 basis_point = 1;
     */
    basisPoint: bigint;
    /**
     * @generated from protobuf field: tdex.v1.Fixed fixed = 2;
     */
    fixed?: Fixed;
}
/**
 * @generated from protobuf message tdex.v1.Fixed
 */
export interface Fixed {
    /**
     * @generated from protobuf field: int64 base_fee = 1;
     */
    baseFee: bigint;
    /**
     * @generated from protobuf field: int64 quote_fee = 2;
     */
    quoteFee: bigint;
}
/**
 * @generated from protobuf message tdex.v1.Balance
 */
export interface Balance {
    /**
     * @generated from protobuf field: uint64 base_amount = 1;
     */
    baseAmount: bigint;
    /**
     * @generated from protobuf field: uint64 quote_amount = 2;
     */
    quoteAmount: bigint;
}
/**
 * @generated from protobuf message tdex.v1.BalanceWithFee
 */
export interface BalanceWithFee {
    /**
     * @generated from protobuf field: tdex.v1.Balance balance = 1;
     */
    balance?: Balance;
    /**
     * @generated from protobuf field: tdex.v1.Fee fee = 2;
     */
    fee?: Fee;
}
/**
 * @generated from protobuf message tdex.v1.Market
 */
export interface Market {
    /**
     * @generated from protobuf field: string base_asset = 1;
     */
    baseAsset: string;
    /**
     * @generated from protobuf field: string quote_asset = 2;
     */
    quoteAsset: string;
}
/**
 * @generated from protobuf message tdex.v1.MarketWithFee
 */
export interface MarketWithFee {
    /**
     * @generated from protobuf field: tdex.v1.Market market = 1;
     */
    market?: Market;
    /**
     * @generated from protobuf field: tdex.v1.Fee fee = 2;
     */
    fee?: Fee;
}
/**
 * @generated from protobuf message tdex.v1.Price
 */
export interface Price {
    /**
     * @generated from protobuf field: double base_price = 1;
     */
    basePrice: number;
    /**
     * @generated from protobuf field: double quote_price = 2;
     */
    quotePrice: number;
}
/**
 * @generated from protobuf message tdex.v1.Preview
 */
export interface Preview {
    /**
     * @generated from protobuf field: tdex.v1.Price price = 1;
     */
    price?: Price;
    /**
     * @generated from protobuf field: tdex.v1.Fee fee = 2;
     */
    fee?: Fee;
    /**
     * @generated from protobuf field: uint64 amount = 3;
     */
    amount: bigint;
    /**
     * @generated from protobuf field: string asset = 4;
     */
    asset: string;
    /**
     * @generated from protobuf field: tdex.v1.Balance balance = 5;
     */
    balance?: Balance;
}
// ...
export const Fee = new Fee$Type();
// ...
export const Fixed = new Fixed$Type();
// ...
export const Balance = new Balance$Type();
// @generated message type with reflection information, may provide speed optimized methods
class BalanceWithFee$Type extends MessageType<BalanceWithFee> {
    constructor() {
        super("tdex.v1.BalanceWithFee", [
            { no: 1, name: "balance", kind: "message", T: () => Balance },
            { no: 2, name: "fee", kind: "message", T: () => Fee }
        ]);
    }
    create(value?: PartialMessage<BalanceWithFee>): BalanceWithFee {
        const message = {};
        globalThis.Object.defineProperty(message, MESSAGE_TYPE, { enumerable: false, value: this });
        if (value !== undefined)
            reflectionMergePartial<BalanceWithFee>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: BalanceWithFee): BalanceWithFee {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* tdex.v1.Balance balance */ 1:
                    message.balance = Balance.internalBinaryRead(reader, reader.uint32(), options, message.balance);
                    break;
                case /* tdex.v1.Fee fee */ 2:
                    message.fee = Fee.internalBinaryRead(reader, reader.uint32(), options, message.fee);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: BalanceWithFee, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* tdex.v1.Balance balance = 1; */
        if (message.balance)
            Balance.internalBinaryWrite(message.balance, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        /* tdex.v1.Fee fee = 2; */
        if (message.fee)
            Fee.internalBinaryWrite(message.fee, writer.tag(2, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message tdex.v1.BalanceWithFee
 */
export const BalanceWithFee = new BalanceWithFee$Type();
// @generated message type with reflection information, may provide speed optimized methods
class Market$Type extends MessageType<Market> {
    constructor() {
        super("tdex.v1.Market", [
            { no: 1, name: "base_asset", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "quote_asset", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<Market>): Market {
        const message = { baseAsset: "", quoteAsset: "" };
        globalThis.Object.defineProperty(message, MESSAGE_TYPE, { enumerable: false, value: this });
        if (value !== undefined)
            reflectionMergePartial<Market>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: Market): Market {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string base_asset */ 1:
                    message.baseAsset = reader.string();
                    break;
                case /* string quote_asset */ 2:
                    message.quoteAsset = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: Market, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string base_asset = 1; */
        if (message.baseAsset !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.baseAsset);
        /* string quote_asset = 2; */
        if (message.quoteAsset !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.quoteAsset);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message tdex.v1.Market
 */
export const Market = new Market$Type();
// ...
export const Price = new Price$Type();
// @generated message type with reflection information, may provide speed optimized methods
class Preview$Type extends MessageType<Preview> {
    constructor() {
        super("tdex.v1.Preview", [
            { no: 1, name: "price", kind: "message", T: () => Price },
            { no: 2, name: "fee", kind: "message", T: () => Fee },
            { no: 3, name: "amount", kind: "scalar", T: 4 /*ScalarType.UINT64*/, L: 0 /*LongType.BIGINT*/ },
            { no: 4, name: "asset", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 5, name: "balance", kind: "message", T: () => Balance }
        ]);
    }
    create(value?: PartialMessage<Preview>): Preview {
        const message = { amount: 0n, asset: "" };
        globalThis.Object.defineProperty(message, MESSAGE_TYPE, { enumerable: false, value: this });
        if (value !== undefined)
            reflectionMergePartial<Preview>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: Preview): Preview {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* tdex.v1.Price price */ 1:
                    message.price = Price.internalBinaryRead(reader, reader.uint32(), options, message.price);
                    break;
                case /* tdex.v1.Fee fee */ 2:
                    message.fee = Fee.internalBinaryRead(reader, reader.uint32(), options, message.fee);
                    break;
                case /* uint64 amount */ 3:
                    message.amount = reader.uint64().toBigInt();
                    break;
                case /* string asset */ 4:
                    message.asset = reader.string();
                    break;
                case /* tdex.v1.Balance balance */ 5:
                    message.balance = Balance.internalBinaryRead(reader, reader.uint32(), options, message.balance);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: Preview, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* tdex.v1.Price price = 1; */
        if (message.price)
            Price.internalBinaryWrite(message.price, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        /* tdex.v1.Fee fee = 2; */
        if (message.fee)
            Fee.internalBinaryWrite(message.fee, writer.tag(2, WireType.LengthDelimited).fork(), options).join();
        /* uint64 amount = 3; */
        if (message.amount !== 0n)
            writer.tag(3, WireType.Varint).uint64(message.amount);
        /* string asset = 4; */
        if (message.asset !== "")
            writer.tag(4, WireType.LengthDelimited).string(message.asset);
        /* tdex.v1.Balance balance = 5; */
        if (message.balance)
            Balance.internalBinaryWrite(message.balance, writer.tag(5, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message tdex.v1.Preview
 */
export const Preview = new Preview$Type();